feat(km-feature-viz): Pillar 3 repivot — ResNet152 + DenseNet121 + GoogLeNet, fp32, 20 images, literature-grounded neuron selection
Refocus the km-feature-viz pipeline for the TMLR Pillar 3 figure
"Exact image-space attribution: knowledge matrices as canonical saliency."

Architecture changes
--------------------
- Drop AlexNet/ResNet18/VGG11 from Pillar 3 (still used by Pillars 1+2).
- Add three new archs: ResNet152, DenseNet121, GoogLeNet (InceptionV1).
- All three are 3x224x224, KM column dim 150 529 — no per-arch shape lookup.
- Library pin bumped to MarcoArmenta/knowledgematrix-cluster@fe64a13ea
  (pillar3 branch stacks: extract_weff + densenet + incep + ResNet152 +
  GoogLeNet wrappers). Verified completeness M(x).sum(1) == f(x) to
  1.4e-5 max abs error on GoogLeNet.

Image budget
------------
- TIER_A_CLASSES = [207 (golden retriever), 282 (tiger cat), 340 (zebra)].
- TIER_A_IMAGES_PER_CLASS = {207: 7, 282: 7, 340: 6} → 20 imgs.
- 20 imgs × 3 archs = 60 manifest entries (was 1500).
- TIER_A_SEED bumped to 20260426.

fp32 storage everywhere
-----------------------
All compute_* scripts store fp32 instead of fp16 — closes the inf-overflow
regression that broke completeness on ResNet18.

Literature-grounded neuron selection (compute_deepdream.py)
-----------------------------------------------------------
Restructured to per-arch methodology:
- ResNet152, DenseNet121: class-conditional Grad-CAM channel weights
  (Selvaraju 2017 channel-wise variant); top-5 channels per layer by mean
  |alpha_y^k| over the 20 manifest images.
- GoogLeNet: catalogued neurons from the Distill Circuits Thread (Olah,
  Cammarata et al. 2020) and OpenAI Microscope. Channel indices preserved
  from the original Caffe model (torchvision's inception4a = Microscope's
  mixed4a). Range-checked at job-start to fail loudly on out-of-range
  catalogue typos.
- Selection persisted to state/03a_neuron_selection_<arch>.json with
  per-channel metadata; renderer reads this for figure captions.
- DeepDream excluded from the headline panel by default
  (--include-deepdream opt-in), per the round-3 Interp/KM debate.

SLURM + orchestrator
--------------------
- job_kmfv_kms.sh, job_kmfv_deepdream.sh: MODELS=("resnet152" "densenet121"
  "googlenet"), --array=0-2.
- job_kmfv_baselines.sh, job_kmfv_formulations.sh: mem 32G→48G for the
  3-arch internal loop.
- job_kmfv_bundle.sh: completeness gate restructured into `counted` (12
  state files at 20/60/15 entries) + `existence_only` (3 neuron-selection
  JSONs).
- run_pipeline.sh: D_MODELS = 3 archs; D1_EXPECTED=20, D2_EXPECTED=60,
  D3_EXPECTED=15; per-arch state-file paths; sbatch dry-run confirmed
  --array=0,1,2 for D1/D3.

Renderer (scripts/render_km_viz.py — newly tracked)
---------------------------------------------------
- ARCHS extended; ARCH_DISPLAY_NAMES added.
- DeepDream caption reads state/03a_neuron_selection_<arch>.json and
  surfaces "<layer> ch{N} (mean alpha=...)" for Grad-CAM-selected or
  "<layer> ch{N} ({label})" for catalogued.
- --include-deepdream flag (default OFF).

Tests
-----
- test_compute_{baselines,kms,deepdream}: parameterized over the 3 archs
  with weights=None for cheap instantiation.
- test_manifest: assert 60 entries.
- New test_neuron_selection.py covers select_neurons_gradcam,
  get_googlenet_catalogued_selection, write_neuron_selection round-trip,
  and range_check_googlenet_neurons (in-range + out-of-range cases).
- 75 tests pass locally (2 deselected as too slow on CPU: smoothgrad and
  IG on resnet152).

+++ km_feature_viz/dictionary.py
@@ -87,13 +87,13 @@ def main() -> int:
         out_components = paths.dictionary_path(model_name, class_id, "components")
         out_var = paths.dictionary_path(model_name, class_id, "explained_variance")
         out_components.parent.mkdir(parents=True, exist_ok=True)
-        torch.save({"pca": pca_comp.to(torch.float16)}, out_components)
+        torch.save({"pca": pca_comp.to(torch.float32)}, out_components)
         torch.save({"pca": pca_var.to(torch.float32)}, out_var)
 
         # NMF
         nmf_comp, nmf_err = nmf_top_k(X, k=args.top_k)
         existing = torch.load(out_components, weights_only=False)
-        existing["nmf"] = nmf_comp.to(torch.float16)
+        existing["nmf"] = nmf_comp.to(torch.float32)
         torch.save(existing, out_components)
         existing_var = torch.load(out_var, weights_only=False)
         existing_var["nmf_err"] = nmf_err
